feat: service단에 검증 예외처리 구현

+++ src/main/java/com/study/petory/domain/tradeBoard/service/TradeBoardServiceImpl.java
@@ -16,6 +16,8 @@ import com.study.petory.domain.tradeBoard.entity.TradeCategory;
 import com.study.petory.domain.tradeBoard.repository.TradeBoardRepository;
 import com.study.petory.domain.user.entity.User;
 import com.study.petory.domain.user.repository.UserRepository;
+import com.study.petory.exception.CustomException;
+import com.study.petory.exception.enums.ErrorCode;
 
 import lombok.RequiredArgsConstructor;
 
@@ -31,7 +33,7 @@ public class TradeBoardServiceImpl implements TradeBoardService {
 	@Transactional
 	public TradeBoardCreateResponseDto saveTradeBoard(TradeBoardCreateRequestDto requestDto) {
 
-		//User user = new User();//나중에 토큰으로 값을 받아올 예정
+		//나중에 토큰으로 값을 받아올 예정
 		User user = userRepository.findById(1L).orElseThrow();
 
 		TradeBoard tradeBoard = TradeBoard.builder()
@@ -57,7 +59,7 @@ public class TradeBoardServiceImpl implements TradeBoardService {
 		PageRequest pageable = PageRequest.of(adjustedPage, 10, Sort.by("createdAt").descending());
 
 		Page<TradeBoard> tradeBoard;
-		if (category != null) {
+		if (category != null) { //카테고리가 있다면 카테고리로 조회
 			tradeBoard = tradeBoardRepository.findAllByCategory(category, pageable);
 		} else {
 			tradeBoard = tradeBoardRepository.findAll(pageable);
@@ -81,7 +83,15 @@ public class TradeBoardServiceImpl implements TradeBoardService {
 	@Transactional
 	public TradeBoardUpdateResponseDto updateTradeBoard(Long tradeBoardId, TradeBoardUpdateRequestDto requestDto) {
 
-		TradeBoard tradeBoard = tradeBoardRepository.findById(tradeBoardId).orElseThrow();
+		//나중에 토큰으로 값 받아오기
+		User user = userRepository.findById(1L).orElseThrow();
+
+		TradeBoard tradeBoard = tradeBoardRepository.findById(tradeBoardId)
+			.orElseThrow(() -> new CustomException(ErrorCode.TRADE_BOARD_NOT_FOUND));
+
+		if (tradeBoard.getUser() != user) {
+			throw new CustomException(ErrorCode.TRADE_BOARD_FORBIDDEN);
+		}
 
 		if (requestDto.getCategory() != null) {
 			tradeBoard.updateCategory(requestDto.getCategory());
@@ -110,7 +120,17 @@ public class TradeBoardServiceImpl implements TradeBoardService {
 	@Override
 	@Transactional
 	public void deleteTradeBoard(Long tradeBoardId) {
-		TradeBoard tradeBoard = tradeBoardRepository.findById(tradeBoardId).orElseThrow();
+
+		//나중에 토큰으로 값 받아오기
+		User user = userRepository.findById(1L).orElseThrow();
+
+		TradeBoard tradeBoard = tradeBoardRepository.findById(tradeBoardId)
+			.orElseThrow(() -> new CustomException(ErrorCode.TRADE_BOARD_NOT_FOUND));
+
+		if (tradeBoard.getUser() != user) {
+			throw new CustomException(ErrorCode.TRADE_BOARD_FORBIDDEN);
+		}
+
 		tradeBoardRepository.delete(tradeBoard);
 	}
 }
